Complete the work for collecting the new fields

+++ trunk/db/migrate/20090721180445_add_all_fields_to_dayusages.rb
@@ -5,18 +5,18 @@ class AddAllFieldsToDayusages < ActiveRecord::Migration
 	add_column :dayusages, :max_pctmem, :float, :default => 0.0
 
 	add_column :dayusages, :avg_phymem, :float, :default => 0.0
-	add_column :dayusages, :total_phymem, :float, :default => 0.0
-	add_column :dayusages, :max_phymem, :float, :default => 0.0
+	add_column :dayusages, :total_phymem, :integer, :default => 0
+	add_column :dayusages, :max_phymem, :integer, :default => 0
 
 	add_column :dayusages, :avg_virmem, :float, :default => 0.0
-	add_column :dayusages, :total_virmem, :float, :default => 0.0
-	add_column :dayusages, :max_virmem, :float, :default => 0.0
+	add_column :dayusages, :total_virmem, :integer, :default => 0
+	add_column :dayusages, :max_virmem, :integer, :default => 0
 
-	add_column :dayusages, :avg_procs, :integer, :default => 0
+	add_column :dayusages, :avg_procs, :float, :default => 0.0
 	add_column :dayusages, :total_procs, :integer, :default => 0
 	add_column :dayusages, :max_procs, :integer, :default => 0
 
-	add_column :dayusages, :avg_runprocs, :integer, :default => 0
+	add_column :dayusages, :avg_runprocs, :float, :default => 0.0
 	add_column :dayusages, :total_runprocs, :integer, :default => 0
 	add_column :dayusages, :max_runprocs, :integer, :default => 0
   end
